+++ CHANGELOG.md
@@ -16,6 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bump `pylint` to `4.0.5`
   ([#4244](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4244))
 
+### Fixed
+
+- `opentelemetry-instrumentation-grpc`: Fix `NotImplementedError` when `grpc.aio` call object does not implement `add_done_callback` (e.g. when a user interceptor awaits the call)
+  ([#4429](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4429))
+
 ### Breaking changes
 
 - Drop Python 3.9 support